watch for sigwinch

+++ src/tui/command_line.rs
@@ -52,7 +52,7 @@ impl CommandLineState {
         self.draw_end_pos.1 - self.cursor_coord.1
     }
 
-    pub async fn get_shell_vars(shell: &ShellClient, width: u16) -> ShellVars {
+    pub async fn get_shell_vars(shell: &ShellClient, width: u32) -> ShellVars {
         shell.do_run(move |shell| {
 
             shell.start_zle_scope();
